requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
google-auth-oauthlib
google-auth
google-api-python-client
python-dotenv
google-generativeai
dotenv